Moved unused component to pages
