--- frontend/src/pages/CreateEvent.jsx
@@ -1,10 +1,0 @@
-import { CreateEventForm } from "../components/CreateEvent/CreateEventForm";
-import PageLayout from "../components/PageLayout";
-
-function CreateEvent() {
-    return(
-    <PageLayout>
-      <CreateEventForm />
-    </PageLayout>
-)};
-export default CreateEvent;
