--- frontend/src/pages/EditEvent.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { EditEventForm } from '../components/CreateEvent/EditEventForm';
-import PageLayout from '../components/PageLayout';
-function EditEventPage() {
-  return (
-    <PageLayout>
-      <EditEventForm />
-    </PageLayout>
-  );
-}
-
-export default EditEventPage;
