--- frontend/src/pages/Login.jsx
@@ -1,6 +1,0 @@
-import LoginForm from "../components/LoginForm"
-
-function Login(){
-    return <LoginForm />
-}
- export default Login
