--- frontend/src/pages/Register.jsx
@@ -1,7 +1,0 @@
-import SignupForm from "../components/RegisterForm"
-
-function Register() {
-    return <SignupForm />
-}
-
-export default Register
